feat(re-formateed controllers): controllers split into sub folders
