+++ server/middlewares/auth.ts
@@ -24,6 +24,7 @@ const authMiddleware = async (req: Request, res: Response, next: NextFunction) =
     });
     if (!foundUser) return res.sendStatus(401);
 
+    console.log(foundUser);
     req.user = foundUser;
     next();
   } catch (err) {
