+++ server/routes/index.ts
@@ -3,10 +3,15 @@ import { Router } from 'express';
 import userRoutes from './user.router';
 // eslint-disable-next-line import/no-unresolved
 import eventRoutes from './event.router';
+// eslint-disable-next-line import/no-unresolved
+import authRoutes from './auth.router';
+// eslint-disable-next-line import/no-unresolved
+// eslint-disable-next-line import/extensions
+import authMiddleware from '../middlewares/auth';
 
 const router = Router();
 
-// Concatting endpoints with user and event routes
+router.use('/auth', authRoutes);
 router.use('/users', userRoutes);
 router.use('/events', eventRoutes);
