+++ server/routes/user.router.ts
@@ -1,65 +1,22 @@
-// Declares file as a module, allows us to declare { Router } in both router files
-// LEGEND :
-// 🚀🚀🚀 = New section
-// ✅ = To be included in MVP
-// 🅱️ = Expect request body
-// 🅿️ = Expect Params (eg. Id)
-// --------------------------------------------------------
-// IMPORTS
 import { Router } from 'express';
 // eslint-disable-next-line import/no-unresolved
 import userController from '../controllers/user.controller';
-import authMiddleware from '../middlewares/auth';
 
 const userRouter = Router();
-// --------------------------------------------------------
-// 🚀🚀🚀 LOGIN + LOGOUT ROUTES 🚀🚀🚀
-// --------------------------------------------------------
 
-// Login 🅱️ ✅
-userRouter.post('/login', userController.login);
-
-// Logout ✅
-userRouter.post('/logout', userController.logout);
-
-// --------------------------------------------------------
-// 🚀🚀🚀 USER ROUTES 🚀🚀🚀
-// --------------------------------------------------------
-// Create 1 user 🅱️ ✅
-userRouter.post('/', userController.createUser);
-
-// Get 1 user by ID 🅿️ ✅
 userRouter.get('/:userid', userController.getUserById);
 
-// Get all users
 userRouter.get('/', userController.getAllUsers);
 
-// Edit 1 user by ID 🅿️ 🅱️
-userRouter.patch('/:userid', authMiddleware, userController.editUserInfo);
+userRouter.patch('/:userid', userController.editUserInfo);
 
-// Get all events by creator ID
 userRouter.get('/:userid/events/created/', userController.getUserCreatedEvents);
 
-// Get all events by creator ID
 userRouter.get('/:userid/events/participating/', userController.getUserParticipatingEvents);
 
-// Delete 1 user by ID 🅿️
 userRouter.delete('/:userid', userController.deleteUser);
 
-// Delete all users (DEV PURPOSES)
+// Dev only
 userRouter.delete('/', userController._deleteAllUsers);
 
-// Add friend 🅿️
-// route and controller Func to be added !
-
-// Remove friend 🅿️
-// route and controller func to be added !
-
-// --------------------------------------------------------
-// 🚀🚀🚀 RATINGS ROUTES 🚀🚀🚀
-// --------------------------------------------------------
-
-// --------------------------------------------------------
-// 🚀🚀🚀 TAGS ROUTES 🚀🚀🚀
-// --------------------------------------------------------
 export default userRouter;
